feat: refactored QuizModalContext, QuizModal and separated QuizList
